+++ src/test/java/com/bitbank/repositories/RolesRepositoryTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import com.bitbank.constants.ERole;
 import com.bitbank.entities.RolesEntity;
 
 @ExtendWith(SpringExtension.class)